+ БД4 лаба + ППС1 лаба + ТРСИС3 лаба + ТП4 слова + ПМО9 лекция + ЭВМ1 отчет

+++ Server_Systems_Development_Technology/Lab_Limenitiz/lab3/src/main/java/suai/trsis2021/lab3/service/CostService.java
@@ -2,6 +2,7 @@ package suai.trsis2021.lab3.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import suai.trsis2021.lab3.entity.CostEntity;
 import suai.trsis2021.lab3.exceptions.CostNotFoundException;
 import suai.trsis2021.lab3.model.Cost;
@@ -9,27 +10,32 @@ import suai.trsis2021.lab3.repository.CostRepository;
 
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
+@Transactional
 public class CostService {
 
     @Autowired
     private CostRepository costRepository;
 
     public boolean addCost(CostEntity cost){
+        if(cost.getName() == null ||
+                cost.getPrice() == null ||
+                cost.getCustomer() == null ||
+                cost.getCategory() == null) {
+            return false;
+        }
+
         try {
-            for (var c : costRepository.findAll()) {
-                if (equalsCosts(c, cost)) {
-                    int count = c.getCount();
-                    costRepository.delete(c);
-                    cost.setCount(count+1);
-                    costRepository.save(cost);
-                    return true;
-                }
+            CostEntity c = findCostEntity(cost);
+            if(c == null){
+                costRepository.save(cost);
             }
-
-
-            costRepository.save(cost);
+            else{
+                c.countUp();
+                updateCostById(c);
+            }
             return true;
         }
         catch (Exception e){
@@ -40,11 +46,46 @@ public class CostService {
 
     public boolean deleteCost(Long id){
         try {
-            costRepository.deleteById(id);
+            Optional<CostEntity> c = costRepository.findById(id);
+            if(c.isEmpty()){
+                return false;
+            }
+
+            CostEntity cost = c.get();
+            if(cost.getCount() > 1){
+                cost.countDown();
+                updateCostById(cost);
+            }
+            else{
+                costRepository.deleteById(id);
+            }
             return true;
+
         }catch (Exception e){
             return false;
         }
+    }
+
+    public boolean putCost(CostEntity cost){
+        if(cost.getName() == null ||
+                cost.getPrice() == null ||
+                cost.getCustomer() == null ||
+                cost.getCategory() == null) {
+            return false;
+        }
+
+        Optional<CostEntity> c = costRepository.findById(cost.getId());
+        if(c.isEmpty()){
+            return false;
+        }
+
+        CostEntity costEntity = c.get();
+        costEntity.setName(cost.getName());
+        costEntity.setPrice(cost.getPrice());
+        costEntity.setCategory(cost.getCategory());
+        costEntity.setCustomer(cost.getCustomer());
+        updateCostById(costEntity);
+        return true;
     }
 
     public List<Cost> getCosts(){
@@ -66,5 +107,25 @@ public class CostService {
                 e1.getCustomer().equals(e2.getCustomer()) &&
                 e1.getPrice().equals(e2.getPrice()) &&
                 e1.getCategory().equals(e2.getCategory());
+    }
+
+    public CostEntity findCostEntity(CostEntity cost) {
+        for (var c : costRepository.findByName(cost.getName())) {
+            if (equalsCosts(c, cost)) {
+                return c;
+            }
+        }
+        return null;
+    }
+
+    private void updateCostById(CostEntity cost){
+        costRepository.updateCostById(
+                cost.getName(),
+                cost.getPrice(),
+                cost.getCategory(),
+                cost.getCustomer(),
+                cost.getCount(),
+                cost.getId()
+        );
     }
 }
